API_Server/requirements.txt
tensorflow
kapre
click
pyyaml
numpy
wavio
faiss-cpu
Flask
tqdm
pandas
pydub
flask-cors